requirements.txt
fastapi
uvicorn
requests
grpcio==1.30.0
grpcio-tools==1.30.0
sec-client/